refactor-code-with-GenericHelper-Class
