+++ AccountantApp/src/test/java/com/tap/project/escrivaghera/AccountantApp/AccountantAppTest.java
@@ -17,7 +17,7 @@ import org.junit.Test;
 
 import com.tap.project.escrivaghera.AccountantApp.exception.IllegalJournalEntryException;
 import com.tap.project.escrivaghera.AccountantApp.exception.NotAuthenticationException;
-import com.tap.project.escrivaghera.AccountantApp.helper.MongoTestHelper;
+import com.tap.project.escrivaghera.AccountantApp.helper.GenericHelper;
 
 public class AccountantAppTest {
 
@@ -48,14 +48,14 @@ public class AccountantAppTest {
 	@Test
 	public void testCreateJournalEntryWithBadListOfCounts() {
 		myAccountantApp.authenticate(user);
-		ArrayList<Count> myCounts = MongoTestHelper.createTestList(1200.0, 1100.0);
+		ArrayList<Count> myCounts = GenericHelper.createTestList(1200.0, 1100.0);
 		assertNull(myAccountantApp.createJournalEntry("1", new Date(), myCounts));
 	}
 
 	@Test
 	public void testCreateJournalEntryWithGoodListOfCounts() {
 		myAccountantApp.authenticate(user);
-		ArrayList<Count> myCounts = MongoTestHelper.createTestList(1200.0, 1200.0);
+		ArrayList<Count> myCounts = GenericHelper.createTestList(1200.0, 1200.0);
 		assertNotNull(myAccountantApp.createJournalEntry("1", new Date(), myCounts));
 	}
 
@@ -90,8 +90,7 @@ public class AccountantAppTest {
 
 	@Test
 	public void testGetAllRegistractionWithOneElement() throws IllegalJournalEntryException {
-		myAccountantApp.authenticate(user);
-		Date[] dates = MongoTestHelper.createDates();
+		Date[] dates = authenticateAndCreateDates();
 		List<JournalEntry> returnList = new ArrayList<JournalEntry>();
 		createJournalEntry(returnList);
 		when(db.getAllRegistration(dates[0], dates[1])).thenReturn(returnList);
@@ -100,8 +99,7 @@ public class AccountantAppTest {
 
 	@Test
 	public void testGetAllRegistractionWithMoreElement() throws IllegalJournalEntryException {
-		myAccountantApp.authenticate(user);
-		Date[] dates = MongoTestHelper.createDates();
+		Date[] dates = authenticateAndCreateDates();
 		List<JournalEntry> returnList = new ArrayList<JournalEntry>();
 		createJournalEntry(returnList);
 		createJournalEntry(returnList);
@@ -142,10 +140,16 @@ public class AccountantAppTest {
 		lengthOfUserActions = myAccountantApp.getMySession().getList().size();
 	}
 
+	private Date[] authenticateAndCreateDates() {
+		myAccountantApp.authenticate(user);
+		Date[] dates = GenericHelper.createDates();
+		return dates;
+	}
+
 	private void createJournalEntry(List<JournalEntry> returnList) throws IllegalJournalEntryException {
 		JournalEntry entry = new JournalEntry("1",
 				new Date(new GregorianCalendar(1900 + 116, 11, 10).getTimeInMillis()));
-		entry.setListOfCount(MongoTestHelper.createTestList(1200.0, 1200.0));
+		entry.setListOfCount(GenericHelper.createTestList(1200.0, 1200.0));
 		returnList.add(entry);
 	}
