+++ AccountantApp/src/test/java/com/tap/project/escrivaghera/AccountantApp/IT/AccountantAppIT.java
@@ -14,8 +14,8 @@ import com.tap.project.escrivaghera.AccountantApp.AccountantApp;
 import com.tap.project.escrivaghera.AccountantApp.JournalEntry;
 import com.tap.project.escrivaghera.AccountantApp.Server;
 import com.tap.project.escrivaghera.AccountantApp.User;
+import com.tap.project.escrivaghera.AccountantApp.helper.GenericHelper;
 import com.tap.project.escrivaghera.AccountantApp.mongo.MongoDatabaseWrapper;
-import com.tap.project.escrivaghera.AccountantApp.helper.MongoTestHelper;
 
 public class AccountantAppIT {
 
@@ -35,22 +35,20 @@ public class AccountantAppIT {
 
 	@Test
 	public void testAddIT() {
-		myAccountantApp.authenticate(myUser);
-		Date[] dates = MongoTestHelper.createDates();
+		Date[] dates = authenticateAndCreateDates();
 		JournalEntry testEntry = myAccountantApp.createJournalEntry("1", dates[0],
-				MongoTestHelper.createTestList(1100, 1100));
+				GenericHelper.createTestList(1100, 1100));
 		myDB.add(testEntry);
 		assertEquals(1, myDB.getAllRegistration(dates[0], dates[1]).size());
 	}
 
 	@Test
 	public void testModifyIT() {
-		myAccountantApp.authenticate(myUser);
-		Date[] dates = MongoTestHelper.createDates();
+		Date[] dates = authenticateAndCreateDates();
 		JournalEntry testEntry1 = myAccountantApp.createJournalEntry("1", dates[0],
-				MongoTestHelper.createTestList(1100, 1100));
+				GenericHelper.createTestList(1100, 1100));
 		JournalEntry testEntry2 = myAccountantApp.createJournalEntry("2", dates[0],
-				MongoTestHelper.createTestList(1200, 1200));
+				GenericHelper.createTestList(1200, 1200));
 		myDB.add(testEntry1);
 		myDB.modify("1", testEntry2);
 		assertEquals(1, myDB.getAllRegistration(dates[0], dates[1]).size());
@@ -58,10 +56,9 @@ public class AccountantAppIT {
 
 	@Test
 	public void testDeleteIT() {
-		myAccountantApp.authenticate(myUser);
-		Date[] dates = MongoTestHelper.createDates();
+		Date[] dates = authenticateAndCreateDates();
 		JournalEntry testEntry = myAccountantApp.createJournalEntry("4", dates[0],
-				MongoTestHelper.createTestList(1300, 1300));
+				GenericHelper.createTestList(1300, 1300));
 		myDB.add(testEntry);
 		myDB.delete("4");
 		assertEquals(0, myDB.getAllRegistration(dates[0], dates[1]).size());
@@ -69,24 +66,22 @@ public class AccountantAppIT {
 
 	@Test
 	public void testGetAllRegistrationITOneElement() {
-		myAccountantApp.authenticate(myUser);
-		Date[] dates = MongoTestHelper.createDates();
+		Date[] dates = authenticateAndCreateDates();
 		JournalEntry testEntry = myAccountantApp.createJournalEntry("1", dates[0],
-				MongoTestHelper.createTestList(1300, 1300));
+				GenericHelper.createTestList(1300, 1300));
 		myDB.add(testEntry);
 		assertEquals(1, myAccountantApp.getAllRegistration(dates[0], dates[1]).size());
 	}
 
 	@Test
 	public void testGetAllRegistrationITMoreElements() {
-		myAccountantApp.authenticate(myUser);
-		Date[] dates = MongoTestHelper.createDates();
+		Date[] dates = authenticateAndCreateDates();
 		JournalEntry testEntry1 = myAccountantApp.createJournalEntry("1", dates[0],
-				MongoTestHelper.createTestList(1300, 1300));
+				GenericHelper.createTestList(1300, 1300));
 		JournalEntry testEntry2 = myAccountantApp.createJournalEntry("2", dates[0],
-				MongoTestHelper.createTestList(1300, 1300));
+				GenericHelper.createTestList(1300, 1300));
 		JournalEntry testEntry3 = myAccountantApp.createJournalEntry("3", dates[0],
-				MongoTestHelper.createTestList(1200, 1200));
+				GenericHelper.createTestList(1200, 1200));
 		myDB.add(testEntry1);
 		myDB.add(testEntry2);
 		myDB.add(testEntry3);
@@ -95,15 +90,14 @@ public class AccountantAppIT {
 
 	@Test
 	public void testGetAllRegistrationITOneBadElement() {
-		myAccountantApp.authenticate(myUser);
-		Date[] dates = MongoTestHelper.createDates();
+		Date[] dates = authenticateAndCreateDates();
 		JournalEntry testEntry1 = myAccountantApp.createJournalEntry("1", dates[0],
-				MongoTestHelper.createTestList(1300, 1300));
+				GenericHelper.createTestList(1300, 1300));
 		JournalEntry testEntry2 = myAccountantApp.createJournalEntry("2",
 				new Date(new GregorianCalendar(1910 + 100, 12, 11).getTimeInMillis()),
-				MongoTestHelper.createTestList(1300, 1300));
+				GenericHelper.createTestList(1300, 1300));
 		JournalEntry testEntry3 = myAccountantApp.createJournalEntry("3", dates[0],
-				MongoTestHelper.createTestList(1200, 1200));
+				GenericHelper.createTestList(1200, 1200));
 		myDB.add(testEntry1);
 		myDB.add(testEntry2);
 		myDB.add(testEntry3);
@@ -112,16 +106,20 @@ public class AccountantAppIT {
 
 	@Test
 	public void testGetAllRegistrationITLostElement() {
-		myAccountantApp.authenticate(myUser);
-		Date[] dates = MongoTestHelper.createDates();
+		Date[] dates = authenticateAndCreateDates();
 		JournalEntry testEntry1 = myAccountantApp.createJournalEntry("1", dates[0],
-				MongoTestHelper.createTestList(1300, 1300));
+				GenericHelper.createTestList(1300, 1300));
 		JournalEntry testEntry2 = myAccountantApp.createJournalEntry("2", dates[0],
-				MongoTestHelper.createTestList(1100, 1100));
+				GenericHelper.createTestList(1100, 1100));
 		myDB.add(testEntry1);
 		myDB.add(testEntry2);
 		myDB.delete("2");
 		assertEquals(1, myAccountantApp.getAllRegistration(dates[0], dates[1]).size());
 	}
 
+	private Date[] authenticateAndCreateDates() {
+		myAccountantApp.authenticate(myUser);
+		Date[] dates = GenericHelper.createDates();
+		return dates;
+	}
 }
