+++ AccountantApp/src/test/java/com/tap/project/escrivaghera/AccountantApp/mongo/MongoDatabaseWrapperTest.java
@@ -1,6 +1,9 @@
 package com.tap.project.escrivaghera.AccountantApp.mongo;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertTrue;
 
 import java.util.Date;
 import java.util.GregorianCalendar;
